+++ pgsui/impute/unsupervised/imputers/ubp.py
@@ -285,6 +285,7 @@ class ImputeUBP(BaseNNImputer):
         self.is_fit_ = True
         self.plotter_.plot_history(self.history_)
         self._evaluate_model(self.X_test_, self.model_, self.best_params_)
+        self._save_best_params(self.best_params_)
         return self
 
     def transform(self) -> np.ndarray: